src/server-new/requirements.txt
django==4.2.7
djangorestframework==3.14.0
btc-lib~=1.0.0